Obs by Encounter and dropped old PatientSummaryActivity
