+++ app/src/main/java/com/muzima/view/forms/FormWebViewActivity.java
@@ -47,7 +47,7 @@ import com.muzima.utils.imaging.ImageResult;
 import com.muzima.utils.video.VideoResult;
 import com.muzima.view.BroadcastListenerActivity;
 import com.muzima.view.barcode.BarcodeCaptureActivity;
-import com.muzima.view.patients.PatientSummaryActivity;
+import com.muzima.view.patients.ClientSummaryActivity;
 import com.muzima.view.progressdialog.MuzimaProgressDialog;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -279,8 +279,8 @@ public class FormWebViewActivity extends BroadcastListenerActivity {
     }
 
     public void startPatientSummaryView(Patient patient) {
-        Intent intent = new Intent(this, PatientSummaryActivity.class);
-        intent.putExtra(PatientSummaryActivity.PATIENT, patient);
+        Intent intent = new Intent(this, ClientSummaryActivity.class);
+        intent.putExtra(ClientSummaryActivity.PATIENT, patient);
         startActivity(intent);
     }
