--- app/src/main/res/layout/item_encounter_observation_2.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright (c) The Trustees of Indiana University, Moi University
-  ~ and Vanderbilt University Medical Center. All Rights Reserved.
-  ~
-  ~ This version of the code is licensed under the MPL 2.0 Open Source license
-  ~ with additional health care disclaimer.
-  ~ If the user is an entity intending to commercialize any application that uses
-  ~  this code in a for-profit venture,please contact the copyright holder.
-  -->
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:layout_height="wrap_content"
-              android:layout_width="fill_parent"
-              android:orientation="vertical">
-    <TextView
-            android:id="@+id/observationHeader"
-            android:fontFamily="@font/roboto_regular"
-            style="@style/ObservationGroupByConceptHeader">
-    </TextView>
-    <TextView
-            android:id="@+id/observationValue"
-            android:fontFamily="@font/montserrat_regular"
-            style="@style/ObservationValue">
-    </TextView>
-
-    <ImageView
-            android:id="@+id/observationComplex"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:contentDescription="@string/complex_obs"
-            android:src="@android:drawable/ic_menu_gallery"
-            android:visibility="gone"/>
-
-    <View android:id="@+id/divider"
-            android:layout_width="match_parent"
-            android:layout_height="1dp"
-            android:layout_marginStart="8dp"
-            android:layout_marginEnd="8dp"
-          android:background="@color/hint_text_grey_opaque"/>
-
-    <TextView
-            android:id="@+id/observationDate"
-            android:fontFamily="@font/roboto_regular"
-            android:textSize="14sp"
-            android:layout_weight=".5"
-            style="@style/ObservationDate">
-    </TextView>
-</LinearLayout>
